+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### 2.1.1
 + [x] Server: fixed uptime positions sorting
++ [x] Server: added config parameter `uptimeUpdateInterval`, default amd minimal value is `5 min` 
 
 ### 2.1.0
 + [x] Server: added interaction with a block height from Mina Explorer